Big Catchup Commit
Been a while. No significant new results in this commit but lots of
smaller tests for various new things in the mean time (RLO, new and
hybrid rates, constant mdot runs, etc). These are all reflected in the
inlists as well. As of this commit, the inlist is setup/commented
accordingly to run a typical "evolve" run with a defined initial age.
RLO, burst counter, and other tested parameters are turned off.